Moved some functions from 'valid' to 'utils' for usage in inference. Now inference creates a directory in the 'store' directory for each track.
