+++ inference.py
@@ -12,111 +12,94 @@ import torch
 import numpy as np
 import soundfile as sf
 import torch.nn as nn
-from utils import prefer_target_instrument
 
 # Using the embedded version of Python can also correctly import the utils module.
 current_dir = os.path.dirname(os.path.abspath(__file__))
 sys.path.append(current_dir)
-from utils import demix, get_model_from_config
+from utils import demix, get_model_from_config, normalize_audio, denormalize_audio, prefer_target_instrument, apply_tta, read_audio_transposed
 
 import warnings
 warnings.filterwarnings("ignore")
 
 
-def run_folder(model, args, config, device, verbose=False):
+def run_folder(model, args, config, device, verbose: bool = False):
+    """
+    Process a folder of audio files for source separation.
+
+    Parameters:
+    ----------
+    model : torch.nn.Module
+        Pre-trained model for source separation.
+    args : Namespace
+        Arguments containing input folder, output folder, and processing options.
+    config : Dict
+        Configuration object with audio and inference settings.
+    device : torch.device
+        Device for model inference (CPU or CUDA).
+    verbose : bool, optional
+        If True, prints detailed information during processing. Default is False.
+    """
+
     start_time = time.time()
     model.eval()
-    all_mixtures_path = glob.glob(args.input_folder + '/*.*')
-    all_mixtures_path.sort()
-    sample_rate = 44100
-    if 'sample_rate' in config.audio:
-        sample_rate = config.audio['sample_rate']
-    print('Total files found: {} Use sample rate: {}'.format(len(all_mixtures_path), sample_rate))
 
-    instruments = prefer_target_instrument(config)[:]
+    mixture_paths = sorted(glob.glob(os.path.join(args.input_folder, '*.*')))
+    sample_rate = config.audio.get('sample_rate', 44100)
 
+    print(f"Total files found: {len(mixture_paths)}. Using sample rate: {sample_rate}")
+
+    instruments = prefer_target_instrument(config)
     os.makedirs(args.store_dir, exist_ok=True)
 
     if not verbose:
-        all_mixtures_path = tqdm(all_mixtures_path, desc="Total progress")
+        mixture_paths = tqdm(mixture_paths, desc="Total progress")
 
     if args.disable_detailed_pbar:
         detailed_pbar = False
     else:
         detailed_pbar = True
 
-    for path in all_mixtures_path:
-        print("Starting processing track: ", path)
-        if not verbose:
-            all_mixtures_path.set_postfix({'track': os.path.basename(path)})
+    for path in mixture_paths:
+        print(f"Processing track: {path}")
         try:
-            mix, sr = librosa.load(path, sr=sample_rate, mono=False)
+            mix, sr = read_audio_transposed(path)
         except Exception as e:
-            print('Cannot read track: {}'.format(path))
-            print('Error message: {}'.format(str(e)))
+            print(f'Cannot read track: {format(path)}')
+            print(f'Error message: {str(e)}')
             continue
 
-        # Convert mono to stereo if needed
-        if len(mix.shape) == 1:
-            mix = np.stack([mix, mix], axis=0)
-
         mix_orig = mix.copy()
-        if 'normalize' in config.inference:
-            if config.inference['normalize'] is True:
-                mono = mix.mean(0)
-                mean = mono.mean()
-                std = mono.std()
-                mix = (mix - mean) / std
+        mix, norm_params = normalize_audio(mix)
+
+        waveforms_orig = demix(config, model, mix, device, model_type=args.model_type, pbar=detailed_pbar)
 
         if args.use_tta:
-            # orig, channel inverse, polarity inverse
-            track_proc_list = [mix.copy(), mix[::-1].copy(), -1. * mix.copy()]
-        else:
-            track_proc_list = [mix.copy()]
+            waveforms_orig = apply_tta(config, model, mix, waveforms_orig, device, args.model_type)
 
-        full_result = []
-        for mix in track_proc_list:
-            waveforms = demix(config, model, mix, device, pbar=detailed_pbar, model_type=args.model_type)
-            full_result.append(waveforms)
-
-        # Average all values in single dict
-        waveforms = full_result[0]
-        for i in range(1, len(full_result)):
-            d = full_result[i]
-            for el in d:
-                if i == 2:
-                    waveforms[el] += -1.0 * d[el]
-                elif i == 1:
-                    waveforms[el] += d[el][::-1].copy()
-                else:
-                    waveforms[el] += d[el]
-        for el in waveforms:
-            waveforms[el] = waveforms[el] / len(full_result)
-
-        # Create a new `instr` in instruments list, 'instrumental' 
         if args.extract_instrumental:
             instr = 'vocals' if 'vocals' in instruments else instruments[0]
+            waveforms_orig['instrumental'] = mix_orig - waveforms_orig[instr]
             if 'instrumental' not in instruments:
                 instruments.append('instrumental')
-            # Output "instrumental", which is an inverse of 'vocals' or the first stem in list if 'vocals' absent
-            waveforms['instrumental'] = mix_orig - waveforms[instr]
+
+        file_name = os.path.splitext(os.path.basename(path))[0]
+
+        output_dir = os.path.join(args.store_dir, file_name)
+        os.makedirs(output_dir, exist_ok=True)
 
         for instr in instruments:
-            estimates = waveforms[instr].T
+            estimates = waveforms_orig[instr]
             if 'normalize' in config.inference:
                 if config.inference['normalize'] is True:
-                    estimates = estimates * std + mean
-            file_name, _ = os.path.splitext(os.path.basename(path))
-            if args.flac_file:
-                output_file = os.path.join(args.store_dir, f"{file_name}_{instr}.flac")
-                subtype = 'PCM_16' if args.pcm_type == 'PCM_16' else 'PCM_24'
-                sf.write(output_file, estimates, sr, subtype=subtype)
-            else:
-                output_file = os.path.join(args.store_dir, f"{file_name}_{instr}.wav")
-                sf.write(output_file, estimates, sr, subtype='FLOAT')
+                    estimates = denormalize_audio(estimates, norm_params)
 
-    time.sleep(1)
-    print("Elapsed time: {:.2f} sec".format(time.time() - start_time))
+            codec = 'flac' if getattr(args, 'flac_file', False) else 'wav'
+            subtype = 'PCM_16' if args.flac_file and args.pcm_type == 'PCM_16' else 'FLOAT'
+
+            output_path = os.path.join(output_dir, f"{instr}.{codec}")
+            sf.write(output_path, estimates.T, sr, subtype=subtype)
+
+    print(f"Elapsed time: {time.time() - start_time:.2f} seconds.")
 
 
 def proc_folder(args):
@@ -143,7 +126,6 @@ def proc_folder(args):
         device = "cpu"
     elif torch.cuda.is_available():
         print('CUDA is available, use --force_cpu to disable it.')
-        device = "cuda"
         device = f'cuda:{args.device_ids[0]}' if type(args.device_ids) == list else f'cuda:{args.device_ids}'
     elif torch.backends.mps.is_available():
         device = "mps"
