+++ valid.py
@@ -14,89 +14,11 @@ import soundfile as sf
 import numpy as np
 import torch.nn as nn
 import multiprocessing
-from utils import demix, get_metrics, get_model_from_config, prefer_target_instrument
+from utils import demix, get_metrics, get_model_from_config, prefer_target_instrument, normalize_audio, denormalize_audio, apply_tta, read_audio_transposed
 from typing import Tuple, Dict, List, Union
 import warnings
 from ml_collections import ConfigDict
 warnings.filterwarnings("ignore")
-
-
-def read_audio_transposed(path: str, instr: str = None, skip_err: bool = False) -> Tuple[np.ndarray, int]:
-    """
-    Reads an audio file, ensuring mono audio is converted to two-dimensional format,
-    and transposes the data to have channels as the first dimension.
-    Parameters
-    ----------
-    path : str
-        Path to the audio file.
-    skip_err: bool
-        If true, not raise errors
-    instr:
-        name of instument
-    Returns
-    -------
-    Tuple[np.ndarray, int]
-        A tuple containing:
-        - Transposed audio data as a NumPy array with shape (channels, length).
-          For mono audio, the shape will be (1, length).
-        - Sampling rate (int), e.g., 44100.
-    """
-
-    try:
-        mix, sr = sf.read(path)
-    except Exception as e:
-        if skip_err:
-            print(f"No stem {instr}: skip!")
-            return None, None
-        else:
-            raise RuntimeError(f"Error reading the file at {path}: {e}")
-    else:
-        if len(mix.shape) == 1:  # For mono audio
-            mix = np.expand_dims(mix, axis=-1)
-        return mix.T, sr
-
-
-def normalize_audio(audio: np.ndarray) -> tuple[np.ndarray, Dict[str, float]]:
-    """
-    Normalize an audio signal by subtracting the mean and dividing by the standard deviation.
-
-    Parameters:
-    ----------
-    audio : np.ndarray
-        Input audio array with shape (channels, time) or (time,).
-
-    Returns:
-    -------
-    tuple[np.ndarray, dict[str, float]]
-        - Normalized audio array with the same shape as the input.
-        - Dictionary containing the mean and standard deviation of the original audio.
-    """
-
-    mono = audio.mean(0)
-    mean, std = mono.mean(), mono.std()
-    return (audio - mean) / std, {"mean": mean, "std": std}
-
-
-def denormalize_audio(audio: np.ndarray, norm_params: Dict[str, float]) -> np.ndarray:
-    """
-    Denormalize an audio signal by reversing the normalization process (multiplying by the standard deviation
-    and adding the mean).
-
-    Parameters:
-    ----------
-    audio : np.ndarray
-        Normalized audio array to be denormalized.
-    norm_params : dict[str, float]
-        Dictionary containing the 'mean' and 'std' values used for normalization.
-
-    Returns:
-    -------
-    np.ndarray
-        Denormalized audio array with the same shape as the input.
-    """
-
-    return audio * norm_params["std"] + norm_params["mean"]
-
 
 
 def logging(logs: List[str], text: str, verbose_logging: bool = False) -> None:
@@ -225,60 +147,6 @@ def update_metrics_and_pbar(
             mixture_paths.set_postfix(pbar_dict)
         except Exception:
             pass
-
-
-def apply_tta(
-        config,
-        model: torch.nn.Module,
-        mix: torch.Tensor,
-        waveforms_orig: Dict[str, torch.Tensor],
-        device: torch.device,
-        model_type: str
-) -> Dict[str, torch.Tensor]:
-    """
-    Apply Test-Time Augmentation (TTA) for source separation.
-
-    This function processes the input mixture with test-time augmentations, including
-    channel inversion and polarity inversion, to enhance the separation results. The
-    results from all augmentations are averaged to produce the final output.
-
-    Parameters:
-    ----------
-    config : Any
-        Configuration object containing model and processing parameters.
-    model : torch.nn.Module
-        The trained model used for source separation.
-    mix : torch.Tensor
-        The mixed audio tensor with shape (channels, time).
-    waveforms_orig : Dict[str, torch.Tensor]
-        Dictionary of original separated waveforms (before TTA) for each instrument.
-    device : torch.device
-        Device (CPU or CUDA) on which the model will be executed.
-    model_type : str
-        Type of the model being used (e.g., "demucs", "custom_model").
-
-    Returns:
-    -------
-    Dict[str, torch.Tensor]
-        Updated dictionary of separated waveforms after applying TTA.
-    """
-    # Create augmentations: channel inversion and polarity inversion
-    track_proc_list = [mix[::-1].copy(), -1.0 * mix.copy()]
-
-    # Process each augmented mixture
-    for i, augmented_mix in enumerate(track_proc_list):
-        waveforms = demix(config, model, augmented_mix, device, model_type=model_type)
-        for el in waveforms:
-            if i == 0:
-                waveforms_orig[el] += waveforms[el][::-1].copy()
-            else:
-                waveforms_orig[el] -= waveforms[el]
-
-    # Average the results across augmentations
-    for el in waveforms_orig:
-        waveforms_orig[el] /= len(track_proc_list) + 1
-
-    return waveforms_orig
 
 
 def process_audio_files(
